requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
colorama==0.4.6
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
